fix(测试跟踪): 取消关联二次提示问题
--bug=1023375 --user=宋昌昌 【测试跟踪】功能用例-编辑-关联测试用例-取消关联无二次提示 https://www.tapd.cn/55049933/s/1338092

+++ test-track/frontend/src/business/case/components/case/CaseTestRelate.vue
@@ -126,6 +126,7 @@ import MsTable from "metersphere-frontend/src/components/new-ui/MsTable";
 import MsTableColumn from "metersphere-frontend/src/components/table/MsTableColumn";
 import TestCaseApiRelate from "@/business/case/components/case/relate/CaseApiRelate";
 import { deleteRelateTest, getRelateTest } from "@/api/testCase";
+import { operationConfirm } from "@/business/utils/sdk-utils";
 import TestCaseScenarioRelate from "@/business/case/components/case/relate/CaseScenarioRelate";
 import TestCaseUiScenarioRelate from "@/business/case/components/case/relate/CaseUiScenarioRelate";
 import TestCaseLoadRelate from "@/business/case/components/case/relate/CaseLoadRelate";
@@ -201,9 +202,15 @@ export default {
       }
     },
     remove(row) {
-      deleteRelateTest(row.testCaseId, row.testId).then(() => {
-        this.$success(this.$t("commons.save_success"), false);
-        this.initTable();
+      this.$confirm(this.$t('case.cancel_relate_case_tips_content'), this.$t('case.cancel_relate_case_tips_title'), {
+        confirmButtonText: this.$t('commons.confirm'),
+        cancelButtonText: this.$t('commons.cancel'),
+        type: 'warning'
+      }).then(() => {
+        deleteRelateTest(row.testCaseId, row.testId).then(() => {
+          this.$success(this.$t("commons.save_success"), false);
+          this.initTable();
+        });
       });
     },
     initTable() {
